+++ android/app/src/main/java/com/example/budgetcircle/database/entities/main/Earnings.kt
@@ -7,7 +7,7 @@ import java.util.*
 
 @Entity (tableName = "earnings")
 data class Earning (
-    val sum: Float,
+    val sum: Double,
     val date: Date,
     val typeId: Int,
     val budgetTypeId: Int
